requirements.txt
aiofiles==23.1.0
blis==0.7.9
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
confection==0.0.4
cupy-cuda116==10.6.0
cymem==2.0.7
fastrlock==0.8.1
filelock==3.9.0
httptools==0.5.0
huggingface-hub==0.12.0
idna==3.4
Jinja2==3.1.2
langcodes==3.3.0
MarkupSafe==2.1.2
multidict==6.0.4
murmurhash==1.0.9
numpy==1.24.2
packaging==23.0
pathy==0.10.1
preshed==3.0.8
pydantic==1.10.4
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
sanic==22.6.2
sanic-routing==22.3.0
six==1.16.0
smart-open==6.3.0
spacy==3.5.0
spacy-alignments==0.9.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
spacy-transformers==1.2.1
srsly==2.4.5
thinc==8.1.7
tokenizers==0.13.2
tqdm==4.64.1
transformers==4.26.1
typer==0.7.0
typing_extensions==4.4.0
urllib3==1.26.14
wasabi==1.1.1
websockets==10.4
en-core-web-trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.5.0/en_core_web_trf-3.5.0-py3-none-any.whl